requirements.txt
fastapi
lxml
pyjwt[crypto]
pymongo
python-dotenv
pytz
requests
requests[socks]
uvicorn[standard]